chore(skip): split out protocol/control.rs
It was getting unwieldy
