+++ qcp/src/protocol/control.rs
@@ -46,25 +46,12 @@
 use std::net::{IpAddr, SocketAddr};
 
 use anyhow::anyhow;
-use figment::{
-    Profile, Provider,
-    value::{Dict, Map},
-};
-use int_enum::IntEnum;
-use num_traits::AsPrimitive;
-use quinn::ConnectionStats;
 use serde::{Deserialize, Serialize};
 use serde_bare::Uint;
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
 use super::common::ProtocolMessage;
-use crate::{
-    Configuration,
-    config::Configuration_Optional,
-    protocol::{DataTag, FindTag as _, TaggedData, Variant, compat::Feature, display_vec_td},
-    transport::ThroughputMode,
-    util::{PortRange as CliPortRange, serialization::SerializeAsString},
-};
+use crate::{protocol::display_vec_td, util::PortRange as CliPortRange};
 
 /// Server banner message, sent on stdout and checked by the client
 pub const BANNER: &str = "qcp-server-2\n";
@@ -78,10 +65,22 @@ pub(crate) const OUR_COMPATIBILITY_NUMERIC: u16 = 3;
 /// The protocol compatibility version implemented by this crate
 pub const OUR_COMPATIBILITY_LEVEL: Compatibility = Compatibility::Level(OUR_COMPATIBILITY_NUMERIC);
 
+mod client_msg;
+pub use client_msg::*;
+
+mod server_msg;
+pub use server_msg::*;
+
+mod greetings;
+pub use greetings::*;
+
+mod closedown;
+pub use closedown::*;
+
 ////////////////////////////////////////////////////////////////////////////////////////
 // Display helpers
 
-use engineering_repr::{EngineeringQuantity as EQ, EngineeringRepr};
+use engineering_repr::EngineeringQuantity as EQ;
 
 fn display_opt_uint(label: &str, bandwidth: Option<&Uint>) -> String {
     bandwidth.map_or_else(String::new, |u| {
@@ -112,7 +111,7 @@ fn display_opt<T: std::fmt::Display>(label: &str, value: Option<&T>) -> String {
 /// * 1: Introduced in qcp 0.3.
 /// * 2: Introduced in qcp 0.5.
 ///
-/// See [`Feature`] for a mapping from compatibility levels to specific features.
+/// See [`crate::protocol::compat::Feature`] for a mapping from compatibility levels to specific features.
 ///
 /// <div class="warning">
 /// While this type implements an automatic `PartialEq`, it does not offer an `Ord` or `PartialOrd`
@@ -214,7 +213,7 @@ impl From<SocketAddr> for ConnectionType {
 
 /// Selects the congestion control algorithm to use.
 /// This structure is serialized as a standard BARE enum.
-/// To serialize it as a string, see [`SerializeAsString`].
+/// To serialize it as a string, see [`crate::util::SerializeAsString`].
 #[derive(
     Copy,
     Clone,
@@ -325,965 +324,6 @@ impl From<PortRange_OnWire> for CliPortRange {
     }
 }
 
-////////////////////////////////////////////////////////////////////////////////////////
-// CLIENT GREETING
-
-#[derive(Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Debug, Default)]
-/// The initial message from client to server.
-///
-/// We have to send this message without knowing what version the server supports.
-pub struct ClientGreeting {
-    /// Protocol compatibility version identifier
-    ///
-    /// This identifies the client's maximum supported protocol sub-version.
-    ///
-    /// N.B. This is not sent as an enum to avoid breaking the server when we have a newer version!
-    pub compatibility: u16,
-    /// Requests the remote emit debug information over the control channel (stderr).
-    pub debug: bool,
-    /// Extension field, reserved for future expansion; for now, must be set to 0
-    pub extension: u8,
-}
-impl ProtocolMessage for ClientGreeting {
-    const WIRE_ENCODING_LIMIT: u32 = 4_096;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-// SERVER GREETING
-
-#[derive(Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Debug, Default)]
-/// The initial message from server to client.
-///
-/// Like [`ClientGreeting`] this is designed to be sent without knowing what version the client supports.
-pub struct ServerGreeting {
-    /// Protocol compatibility version identifier
-    ///
-    /// This identifies the client's maximum supported protocol sub-version.
-    ///
-    /// N.B. This is not sent as an enum to avoid breaking the server when we have a newer version!
-    pub compatibility: u16,
-    /// Extension field, reserved for future expansion; for now, must be set to 0
-    pub extension: u8,
-}
-impl ProtocolMessage for ServerGreeting {
-    const WIRE_ENCODING_LIMIT: u32 = 4_096;
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-// CLIENT MESSAGE
-
-#[derive(
-    Clone,
-    Serialize,
-    Deserialize,
-    PartialEq,
-    Debug,
-    Default,
-    derive_more::Display,
-    derive_more::From,
-)]
-/// The control parameters send from client to server.
-pub enum ClientMessage {
-    /// Special value indicating an endpoint has not yet read the remote `ClientMessage`.
-    /// This value should never be seen on the wire.
-    #[default]
-    #[serde(skip_serializing)]
-    ToFollow, // 0
-    /// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-    /// On the wire this is encoded with enum discriminant 1.
-    V1(ClientMessageV1),
-    /// This version was introduced with `VersionCompatibility` level 3.
-    /// On the wire this is encoded with enum discriminant 2.
-    V2(ClientMessageV2),
-}
-impl ProtocolMessage for ClientMessage {}
-
-#[derive(
-    Clone, Serialize, Deserialize, PartialEq, Default, derive_more::Debug, derive_more::Display,
-)]
-// We define a complicated Display here for efficiency; omit fields which are None.
-#[display(
-    "{connection_type}{}{}{}{}{}{}{}, attributes {}",
-    display_opt("remote port", port.as_ref()),
-    display_opt_uint("bw to client", bandwidth_to_client.as_ref()),
-    display_opt_uint("bw to server", bandwidth_to_server.as_ref()),
-    display_opt("RTT", rtt.as_ref()),
-    display_opt("congestion algorithm ", congestion.as_ref()),
-    display_opt_uint("cwnd ", initial_congestion_window.as_ref()),
-    display_opt("timeout", timeout.as_ref()),
-    display_vec_td(attributes),
-)]
-/// Version 1 of the client control parameters message.
-/// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-pub struct ClientMessageV1 {
-    /// Client's self-signed certificate (DER)
-    #[debug(ignore)]
-    pub cert: Vec<u8>,
-    /// The connection type to use (the type of socket we want the server to bind)
-    pub connection_type: ConnectionType,
-    /// If present, requests the server bind to a UDP port from a given range.
-    pub port: Option<PortRange_OnWire>,
-    /// Requests the server show its configuration for this connection
-    pub show_config: bool,
-
-    /// The requested bandwidth to use from client to server
-    pub bandwidth_to_server: Option<Uint>,
-    /// The requested bandwidth to use from server to client (if None, use the same as bandwidth to server)
-    pub bandwidth_to_client: Option<Uint>,
-    /// The network Round Trip Time, in milliseconds, to use in calculating the bandwidth delay product
-    pub rtt: Option<u16>,
-    /// The congestion control algorithm to use
-    pub congestion: Option<CongestionController>,
-    /// The initial congestion window, if specified
-    pub initial_congestion_window: Option<Uint>,
-    /// Connection timeout for the QUIC endpoints, in seconds
-    pub timeout: Option<u16>,
-
-    /// Optional extended attributes
-    ///
-    /// If it is mandatory for the server to action a given attribute, it MUST NOT be sent in this field.
-    /// Instead, use a later version of the `ClientMessage`.
-    ///
-    /// This field was added in qcp 0.5 with `VersionCompatibility` level 2.
-    /// Prior to Compatibility::Level(2) this was a reserved u8, which was required to be set to 0.
-    /// If length 0, it looks the same on the wire.
-    /// If length >0, earlier versions ignore the attributes.
-    pub attributes: Vec<TaggedData<ClientMessageAttributes>>,
-}
-
-/// Extensible credentials tag
-///
-/// This enum was introduced with `VersionCompatibility` level 3.
-#[derive(
-    strum_macros::Display,
-    Clone,
-    Copy,
-    Debug,
-    Default,
-    IntEnum,
-    PartialEq,
-    Serialize,
-    Deserialize,
-    clap::ValueEnum,
-    strum_macros::EnumString,
-    strum_macros::VariantNames,
-)]
-#[non_exhaustive]
-#[repr(u64)]
-#[strum(serialize_all = "lowercase")] // N.B. this applies to EnumString, not Display
-#[serde(rename_all = "lowercase")]
-pub enum CredentialsType {
-    /// Indicates an invalid attribute, or that there is no preference.
-    /// This enum variant is not valid on the wire; credentials must always be a concrete type.
-    Any = 0,
-    /// Self-signed X509 certificate.
-    // Data is a `Variant::Bytes`
-    X509,
-    /// Raw (RFC7250) public key.
-    // Data is a `Variant::Bytes`
-    #[default]
-    RawPublicKey,
-}
-impl DataTag for CredentialsType {}
-
-#[derive(
-    Clone, Default, Serialize, Deserialize, PartialEq, derive_more::Debug, derive_more::Display,
-)]
-#[display("{connection_type}, attributes {}", display_vec_td(attributes))]
-/// Version 2 of the client control parameters message.
-/// This version was introduced with `VersionCompatibility` level 3.
-pub struct ClientMessageV2 {
-    /// Client's TLS credentials (DER)
-    #[debug(ignore)]
-    pub credentials: TaggedData<CredentialsType>,
-
-    /// The connection type to use (the type of socket we want the server to bind)
-    pub connection_type: ConnectionType,
-
-    /// Optional fields
-    pub attributes: Vec<TaggedData<ClientMessage2Attributes>>,
-
-    /// Extension field, reserved for future expansion; for now, must be set to 0
-    pub extension: u8,
-}
-
-impl ClientMessageV2 {
-    fn new(credentials: TaggedData<CredentialsType>, connection_type: ConnectionType) -> Self {
-        Self {
-            credentials,
-            connection_type,
-            attributes: Vec::new(),
-            extension: 0,
-        }
-    }
-}
-
-impl ClientMessageV2 {
-    pub(crate) fn apply_config_attributes(
-        &mut self,
-        remote_config: bool,
-        our_config: &Configuration_Optional,
-    ) {
-        if remote_config {
-            self.attributes
-                .push(ClientMessage2Attributes::OutputConfig.into());
-        }
-        if let Some(pr) = our_config.remote_port {
-            self.attributes
-                .push(ClientMessage2Attributes::PortRangeStart.with_unsigned(pr.begin));
-            self.attributes
-                .push(ClientMessage2Attributes::PortRangeEnd.with_unsigned(pr.end));
-        }
-        if let Some(eq) = our_config.tx {
-            self.attributes
-                .push(ClientMessage2Attributes::BandwidthToServer.with_unsigned(u64::from(eq)));
-        }
-        if let Some(eq) = our_config.rx {
-            self.attributes
-                .push(ClientMessage2Attributes::BandwidthToClient.with_unsigned(u64::from(eq)));
-        }
-        if let Some(rtt) = our_config.rtt {
-            self.attributes
-                .push(ClientMessage2Attributes::RoundTripTime.with_unsigned(rtt));
-        }
-        if let Some(cc) = our_config.congestion {
-            self.attributes
-                .push(ClientMessage2Attributes::CongestionControllerType.with_unsigned(*cc as u64));
-        }
-        if let Some(icw) = our_config.initial_congestion_window {
-            self.attributes.push(
-                ClientMessage2Attributes::InitialCongestionWindow.with_unsigned(u64::from(icw)),
-            );
-        }
-        if let Some(t) = our_config.timeout {
-            self.attributes
-                .push(ClientMessage2Attributes::QuicTimeout.with_unsigned(t));
-        }
-        // DirectionOfTravel is set up by set_direction()
-    }
-}
-
-impl From<ClientMessageV1> for ClientMessageV2 {
-    fn from(v1: ClientMessageV1) -> Self {
-        let mut attributes = Vec::new();
-        if let Some(pr) = v1.port {
-            attributes.push(ClientMessage2Attributes::PortRangeStart.with_unsigned(pr.begin));
-            attributes.push(ClientMessage2Attributes::PortRangeEnd.with_unsigned(pr.end));
-        }
-        if v1.show_config {
-            attributes.push(ClientMessage2Attributes::OutputConfig.into());
-        }
-        if let Some(Uint(bw)) = v1.bandwidth_to_server {
-            attributes.push(ClientMessage2Attributes::BandwidthToServer.with_unsigned(bw));
-        }
-        if let Some(Uint(bw)) = v1.bandwidth_to_client {
-            attributes.push(ClientMessage2Attributes::BandwidthToClient.with_unsigned(bw));
-        }
-        if let Some(rtt) = v1.rtt {
-            attributes.push(ClientMessage2Attributes::RoundTripTime.with_unsigned(rtt));
-        }
-        if let Some(cc) = v1.congestion {
-            attributes
-                .push(ClientMessage2Attributes::CongestionControllerType.with_unsigned(cc as u64));
-        }
-        if let Some(Uint(icw)) = v1.initial_congestion_window {
-            attributes.push(ClientMessage2Attributes::InitialCongestionWindow.with_unsigned(icw));
-        }
-        if let Some(t) = v1.timeout {
-            attributes.push(ClientMessage2Attributes::QuicTimeout.with_unsigned(t));
-        }
-        if let Some(v) = v1
-            .attributes
-            .find_tag(ClientMessageAttributes::DirectionOfTravel)
-        {
-            attributes.push(ClientMessage2Attributes::DirectionOfTravel.with_variant(v.clone()));
-        }
-        Self {
-            credentials: CredentialsType::X509.with_variant(v1.cert.into()),
-            connection_type: v1.connection_type,
-            attributes,
-            extension: 0,
-        }
-    }
-}
-
-#[cfg(test)]
-#[derive(Clone, Serialize, Deserialize, PartialEq, Debug, Default)]
-/// The control parameters send from client to server.
-enum OriginalClientMessage {
-    #[default]
-    #[serde(skip_serializing)]
-    ToFollow,
-    V1(OriginalClientMessageV1),
-}
-#[cfg(test)]
-impl ProtocolMessage for OriginalClientMessage {}
-
-#[cfg(test)]
-#[derive(Clone, Serialize, Deserialize, PartialEq, Default, derive_more::Debug)]
-struct OriginalClientMessageV1 {
-    cert: Vec<u8>,
-    connection_type: ConnectionType,
-    port: Option<PortRange_OnWire>,
-    show_config: bool,
-    bandwidth_to_server: Option<Uint>,
-    bandwidth_to_client: Option<Uint>,
-    rtt: Option<u16>,
-    congestion: Option<CongestionController>,
-    initial_congestion_window: Option<Uint>,
-    timeout: Option<u16>,
-    extension: u8,
-}
-
-/// Extension attributes for [`ClientMessageV1`]
-///
-/// This enum was introduced in qcp 0.5 with `VersionCompatibility` level 2.
-#[derive(strum_macros::Display, Clone, Copy, Debug, IntEnum, PartialEq)]
-#[non_exhaustive]
-#[repr(u64)]
-pub enum ClientMessageAttributes {
-    /// Indicates an invalid attribute.
-    Invalid = 0,
-    /// The intended direction of data flow for the connection.
-    /// This is a value from [`Direction`], stored as [`crate::protocol::Variant::Unsigned`].
-    DirectionOfTravel,
-}
-impl DataTag for ClientMessageAttributes {
-    fn debug_data(&self, data: &Variant) -> String {
-        match self {
-            ClientMessageAttributes::DirectionOfTravel => {
-                Direction::from_repr(data.coerce_unsigned().as_())
-                    .unwrap_or(Direction::Both)
-                    .to_string()
-            }
-            _ => format!("{data:?}"),
-        }
-    }
-}
-
-/// Extension attributes for `ClientMessageV2`
-///
-/// This enum was introduced with `VersionCompatibility` level 3.
-#[derive(strum_macros::Display, Clone, Copy, Debug, IntEnum, PartialEq)]
-#[non_exhaustive]
-#[repr(u64)]
-pub enum ClientMessage2Attributes {
-    /// Indicates an invalid attribute.
-    Invalid = 0,
-    /// The intended direction of data flow for the connection.
-    /// This is a value from [`Direction`], stored as [`crate::protocol::Variant::Unsigned`].
-    DirectionOfTravel,
-    /// Specifies the start of the port range we would like the server
-    /// to bind to.
-    /// Must always be used with `PortRangeEnd`.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    PortRangeStart,
-    /// Specifies the end of the port range we would like the server
-    /// to bind to.
-    /// Must always be used with `PortRangeStart`.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    PortRangeEnd,
-    /// Requests the server to output its config for this connection.
-    /// Data is Empty.
-    OutputConfig,
-    /// The requested bandwidth to use from client to server.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    BandwidthToServer,
-    /// The requested bandwidth to use from server to client (if None, use the same as bandwidth to server).
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    BandwidthToClient,
-    /// The network Round Trip Time, in milliseconds, to use in calculating the bandwidth delay product.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    RoundTripTime,
-    /// The congestion control algorithm to use.
-    /// This is a value from [`CongestionController`], stored as [`crate::protocol::Variant::Unsigned`].
-    CongestionControllerType,
-    /// The initial congestion window.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    InitialCongestionWindow,
-    /// Connection timeout for the QUIC endpoints, in seconds.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    QuicTimeout,
-}
-impl DataTag for ClientMessage2Attributes {
-    fn debug_data(&self, data: &Variant) -> String {
-        match self {
-            ClientMessage2Attributes::DirectionOfTravel => {
-                Direction::from_repr(data.coerce_unsigned().as_())
-                    .unwrap_or(Direction::Both)
-                    .to_string()
-            }
-            ClientMessage2Attributes::CongestionControllerType => {
-                CongestionController::from_repr(data.coerce_unsigned().as_())
-                    .unwrap_or_default()
-                    .to_string()
-            }
-            ClientMessage2Attributes::BandwidthToClient
-            | ClientMessage2Attributes::BandwidthToServer => {
-                data.coerce_unsigned().to_eng(4).to_string()
-            }
-            _ => format!("{data:?}"),
-        }
-    }
-}
-
-/// Direction of data flow for the connection.
-///
-/// This enum was introduced in qcp 0.5 with `VersionCompatibility` level 2.
-#[derive(
-    strum_macros::Display, Clone, Copy, Debug, PartialEq, Eq, strum_macros::FromRepr, Default,
-)]
-#[allow(missing_docs)]
-pub enum Direction {
-    #[default]
-    Both,
-    ClientToServer,
-    ServerToClient,
-}
-impl From<Direction> for Variant {
-    fn from(value: Direction) -> Self {
-        Variant::unsigned(value as u64)
-    }
-}
-impl From<&Variant> for Direction {
-    /// An infallible, type-coercing conversion.
-    /// If the Variant is an unexpected type, returns the default (`Both`).
-    fn from(value: &Variant) -> Self {
-        Direction::from_repr(value.coerce_unsigned().as_()).unwrap_or_default()
-    }
-}
-impl From<Option<&Variant>> for Direction {
-    /// An infallible, type-coercing conversion.
-    /// If the Variant is an unexpected type, returns the default (`Both`).
-    fn from(value: Option<&Variant>) -> Self {
-        value.map_or(Direction::default(), Direction::from)
-    }
-}
-impl Direction {
-    pub(crate) fn server_mode(self) -> ThroughputMode {
-        match self {
-            Direction::ClientToServer => ThroughputMode::Rx,
-            Direction::ServerToClient => ThroughputMode::Tx,
-            Direction::Both => ThroughputMode::Both,
-        }
-    }
-    pub(crate) fn client_mode(self) -> ThroughputMode {
-        match self {
-            Direction::ClientToServer => ThroughputMode::Tx,
-            Direction::ServerToClient => ThroughputMode::Rx,
-            Direction::Both => ThroughputMode::Both,
-        }
-    }
-}
-
-impl ClientMessage {
-    pub(crate) fn new(
-        compat: Compatibility,
-        cert: TaggedData<CredentialsType>,
-        connection_type: ConnectionType,
-        remote_config: bool,
-        my_config: &Configuration_Optional,
-    ) -> Self {
-        assert!(cert.data.is_bytes());
-        if compat.supports(Feature::CMSG_SMSG_2) {
-            let mut msg = ClientMessageV2::new(cert, connection_type);
-            msg.apply_config_attributes(remote_config, my_config);
-            msg.into()
-        } else {
-            let cert_bytes = cert.data.into_bytes().unwrap_or_default();
-            ClientMessageV1::new(&cert_bytes, connection_type, remote_config, my_config).into()
-        }
-    }
-
-    pub(crate) fn set_direction(&mut self, direction: Direction) {
-        match self {
-            ClientMessage::ToFollow => (),
-            ClientMessage::V1(msg) => msg
-                .attributes
-                .push(ClientMessageAttributes::DirectionOfTravel.with_unsigned(direction as u64)),
-            ClientMessage::V2(msg) => msg
-                .attributes
-                .push(ClientMessage2Attributes::DirectionOfTravel.with_unsigned(direction as u64)),
-        }
-    }
-}
-
-impl ClientMessageV1 {
-    fn new(
-        cert: &[u8],
-        connection_type: ConnectionType,
-        remote_config: bool,
-        my_config: &Configuration_Optional,
-    ) -> Self {
-        use engineering_repr::EngineeringQuantity as EQ;
-        // Configuration_Optional seems a bit much for recent rust-analyzer, but the compiler doesn't mind it. Sop:
-        let rx: &Option<EQ<u64>> = &my_config.rx;
-        let icw: &Option<EQ<u64>> = &my_config.initial_congestion_window;
-
-        Self {
-            cert: cert.to_vec(),
-            connection_type,
-            port: my_config.remote_port.map(std::convert::Into::into),
-            show_config: remote_config,
-
-            bandwidth_to_server: match my_config.tx.map(u64::from) {
-                None | Some(0) => None,
-                Some(v) => Some(Uint(v)),
-            },
-            bandwidth_to_client: rx.map(|u| Uint(u64::from(u))),
-            rtt: my_config.rtt,
-            congestion: my_config
-                .congestion
-                .map(|o: SerializeAsString<CongestionController>| *o),
-            initial_congestion_window: icw.map(|u| Uint(u64::from(u))),
-            timeout: my_config.timeout,
-
-            attributes: vec![],
-        }
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-// SERVER MESSAGE
-
-#[derive(
-    Clone,
-    Serialize,
-    Default,
-    Deserialize,
-    PartialEq,
-    Debug,
-    derive_more::From,
-    derive_more::Display,
-)]
-/// The control parameters sent from server to client
-pub enum ServerMessage {
-    /// Special value indicating an endpoint has not yet read the remote `ServerMessage`.
-    /// This value should never be seen on the wire.
-    #[default]
-    #[serde(skip_serializing)]
-    ToFollow,
-    /// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-    /// On the wire enum discriminant: 1.
-    V1(ServerMessageV1),
-    /// This message type was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-    /// On the wire enum discriminant: 2.
-    Failure(ServerFailure),
-    /// This message type was introduced with `VersionCompatibility` level 3.
-    /// On the wire enum discriminant: 3.
-    V2(ServerMessageV2),
-}
-impl ProtocolMessage for ServerMessage {}
-
-impl ServerMessage {
-    pub(crate) fn new(
-        compat: Compatibility,
-        config: &Configuration,
-        port: u16,
-        credentials: TaggedData<CredentialsType>,
-        common_name: String,
-        warning: String,
-    ) -> Self {
-        assert!(credentials.data.is_bytes());
-        let bandwidth_to_server = Uint(config.rx());
-        let bandwidth_to_client = Uint(config.tx());
-        if compat.supports(Feature::CMSG_SMSG_2) {
-            let mut msg = ServerMessageV2 {
-                port,
-                credentials,
-                common_name,
-                bandwidth_to_server,
-                bandwidth_to_client,
-                rtt: config.rtt,
-                ..Default::default()
-            };
-            msg.apply_config_attributes(config);
-            msg.into()
-        } else {
-            let cert_bytes = credentials.data.into_bytes().unwrap_or_default();
-            ServerMessageV1 {
-                port,
-                cert: cert_bytes,
-                name: common_name,
-                bandwidth_to_server,
-                bandwidth_to_client,
-                rtt: config.rtt,
-                congestion: *config.congestion,
-                initial_congestion_window: Uint(config.initial_congestion_window.into()),
-                timeout: config.timeout,
-                warning,
-                ..Default::default()
-            }
-            .into()
-        }
-    }
-}
-
-#[derive(
-    Clone, Serialize, Deserialize, PartialEq, Eq, derive_more::Debug, Default, derive_more::Display,
-)]
-/// Version 1 of the message from server to client.
-/// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-#[display(
-    "{name}:{port} in {}, out {}, rtt {rtt}, congestion {congestion}/{}, timeout {timeout}, \"{warning}\"",
-    bandwidth_to_server.0,
-    bandwidth_to_client.0,
-    initial_congestion_window.0,
-)]
-pub struct ServerMessageV1 {
-    /// UDP port the server has bound to
-    pub port: u16,
-    /// Server's self-signed certificate (DER)
-    #[debug(ignore)]
-    pub cert: Vec<u8>,
-    /// Name in the server cert (this saves us having to unpick it from the certificate)
-    pub name: String,
-
-    /// The final bandwidth to use from client to server
-    pub bandwidth_to_server: Uint,
-    /// The final bandwidth to use from server to client
-    pub bandwidth_to_client: Uint,
-    /// The final round-trip-time to use on the connection
-    pub rtt: u16,
-    /// The congestion control algorithm to use
-    pub congestion: CongestionController,
-    /// The initial congestion window to use (0 means "use algorithm default")
-    pub initial_congestion_window: Uint,
-    /// Connection timeout for the QUIC endpoints, in seconds
-    pub timeout: u16,
-
-    /// If non-zero length, this is a warning message to be relayed to a human
-    pub warning: String,
-    /// Extension field, reserved for future expansion; for now, must be set to 0
-    pub extension: u8,
-}
-
-impl ServerMessageV1 {
-    const META_NAME: &str = "server message";
-}
-
-impl Provider for ServerMessageV1 {
-    fn metadata(&self) -> figment::Metadata {
-        figment::Metadata::named(Self::META_NAME)
-    }
-
-    fn data(&self) -> Result<figment::value::Map<figment::Profile, Dict>, figment::Error> {
-        let mut dict = Dict::new();
-        let mut insert = |key: &str, val: figment::value::Value| {
-            let _ = dict.insert(key.into(), val);
-        };
-        // This is written from the consumer's (client's) point of view, i.e. bandwidth_to_server is client's tx.
-        insert("tx", self.bandwidth_to_server.0.into());
-        insert("rx", self.bandwidth_to_client.0.into());
-
-        insert("rtt", self.rtt.into());
-        insert("congestion", self.congestion.to_string().into());
-        insert("timeout", self.timeout.into());
-        insert(
-            "initial_congestion_window",
-            self.initial_congestion_window.0.into(),
-        );
-
-        let mut profile_map = Map::new();
-        let _ = profile_map.insert(Profile::Global, dict);
-
-        Ok(profile_map)
-    }
-}
-
-/// A special type of message indicating that an error occurred and the connection cannot proceed.
-///
-/// Protocol Version Compatibility: V1
-#[derive(Clone, Serialize, Deserialize, PartialEq, Eq, Debug, derive_more::Display)]
-pub enum ServerFailure {
-    /// The server failed to understand control channel traffic received from the client.
-    ///
-    /// Protocol Version Compatibility: V1
-    Malformed,
-    /// The client's configuration and server's configuration could not be reconciled.
-    /// The string within explains why.
-    ///
-    /// Protocol Version Compatibility: V1
-    #[display("Negotiation Failed: {_0}")]
-    NegotiationFailed(String),
-    /// The QUIC endpoint could not be set up.
-    /// The string within contains more detail.
-    ///
-    /// Protocol Version Compatibility: V1
-    #[display("Endpoint Failed: {_0}")]
-    EndpointFailed(String),
-    /// An unknown error occurred. This is a catch-all for forward compatibility.
-    ///
-    /// Protocol Version Compatibility: V1
-    #[display("Unknown error: {_0}")]
-    Unknown(String),
-}
-
-/// Version 2 of the server control parameters message.
-/// This version was introduced with `VersionCompatibility` level 3.
-#[derive(
-    Clone, Serialize, Default, Deserialize, PartialEq, derive_more::Debug, derive_more::Display,
-)]
-#[display(
-    "{common_name}:{port} in {}, out {}, rtt {rtt}, attrs {}",
-    bandwidth_to_server.0.to_eng(4),
-    bandwidth_to_client.0.to_eng(4),
-    display_vec_td(attributes)
-)]
-pub struct ServerMessageV2 {
-    /// UDP port the server has bound to
-    pub port: u16,
-    /// Server's TLS credentials (DER encoded)
-    #[debug(ignore)]
-    pub credentials: TaggedData<CredentialsType>,
-
-    /// Server's common name, if provided in the credentials.
-    /// This saves us having to unpick it from the certificate.
-    pub common_name: String,
-
-    /// The final bandwidth to use from client to server
-    pub bandwidth_to_server: Uint,
-    /// The final bandwidth to use from server to client
-    pub bandwidth_to_client: Uint,
-    /// The final round-trip-time to use on the connection
-    pub rtt: u16,
-
-    /// Optional fields
-    pub attributes: Vec<TaggedData<ServerMessage2Attributes>>,
-    // OPTIONAL: name?
-    /// Extension field, reserved for future expansion; for now, must be set to 0
-    pub extension: u8,
-}
-
-impl From<ServerMessageV1> for ServerMessageV2 {
-    fn from(v1: ServerMessageV1) -> Self {
-        let mut attributes = Vec::new();
-        if !v1.warning.is_empty() {
-            attributes
-                .push(ServerMessage2Attributes::WarningMessage.with_variant(v1.warning.into()));
-        }
-        attributes.push(
-            ServerMessage2Attributes::CongestionController.with_unsigned(v1.congestion as u64),
-        );
-        if v1.initial_congestion_window.0 != 0 {
-            attributes.push(
-                ServerMessage2Attributes::InitialCongestionWindow
-                    .with_unsigned(v1.initial_congestion_window.0),
-            );
-        }
-        if v1.timeout != 0 {
-            attributes.push(ServerMessage2Attributes::QuicTimeout.with_unsigned(v1.timeout));
-        }
-        Self {
-            port: v1.port,
-            credentials: CredentialsType::X509.with_variant(v1.cert.into()),
-            common_name: v1.name,
-            bandwidth_to_server: v1.bandwidth_to_server,
-            bandwidth_to_client: v1.bandwidth_to_client,
-            rtt: v1.rtt,
-            attributes,
-            extension: 0,
-        }
-    }
-}
-
-/// Optional attributes for `ServerMessageV2`
-///
-/// This enum was introduced with `VersionCompatibility` level 3.
-#[derive(strum_macros::Display, Clone, Copy, Debug, IntEnum, PartialEq)]
-#[non_exhaustive]
-#[repr(u64)]
-pub enum ServerMessage2Attributes {
-    /// Indicates an invalid attribute.
-    Invalid = 0,
-
-    /// The congestion control algorithm to use.
-    /// This is a value from [`CongestionController`], stored as [`crate::protocol::Variant::Unsigned`].
-    ///
-    /// The [`ServerMessage2Attributes::InitialCongestionWindow`] attribute may be used to specify
-    /// the initial congestion window, if required.
-    CongestionController,
-
-    /// The initial congestion window to use. If not present, the algorithm default is used.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    InitialCongestionWindow,
-
-    /// A warning message to be relayed to a human.
-    /// Data is [`crate::protocol::Variant::String`].
-    WarningMessage,
-
-    /// Connection timeout for the QUIC endpoints, in seconds.
-    /// Data is [`crate::protocol::Variant::Unsigned`].
-    QuicTimeout,
-}
-
-impl DataTag for ServerMessage2Attributes {}
-
-impl ServerMessageV2 {
-    pub(crate) fn apply_config_attributes(&mut self, config: &Configuration) {
-        if *config.congestion != CongestionController::default() {
-            self.attributes.push(
-                ServerMessage2Attributes::CongestionController
-                    .with_unsigned(*config.congestion as u64),
-            );
-        }
-        let window = u64::from(config.initial_congestion_window);
-        if window != 0 {
-            self.attributes
-                .push(ServerMessage2Attributes::InitialCongestionWindow.with_unsigned(window));
-        }
-        if config.timeout != 0 {
-            self.attributes
-                .push(ServerMessage2Attributes::QuicTimeout.with_unsigned(config.timeout));
-        }
-        // WarningMessage is set up when the message is created.
-    }
-}
-
-impl Provider for ServerMessageV2 {
-    fn metadata(&self) -> figment::Metadata {
-        figment::Metadata::named("ServerMessageV2")
-    }
-
-    fn data(&self) -> Result<figment::value::Map<figment::Profile, Dict>, figment::Error> {
-        let mut dict = Dict::new();
-        let mut insert = |key: &str, val: figment::value::Value| {
-            let _ = dict.insert(key.into(), val);
-        };
-        // This is written from the consumer's (client's) point of view, i.e. bandwidth_to_server is client's tx.
-        insert("tx", self.bandwidth_to_server.0.into());
-        insert("rx", self.bandwidth_to_client.0.into());
-        insert("rtt", self.rtt.into());
-
-        for attr in &self.attributes {
-            if let Some(tag) = attr.tag() {
-                let data = &attr.data;
-                match tag {
-                    ServerMessage2Attributes::CongestionController => {
-                        let ctrl = data.coerce_unsigned().as_();
-                        let cc = CongestionController::from_repr(ctrl).unwrap_or_default();
-                        insert("congestion", cc.into());
-                    }
-                    ServerMessage2Attributes::InitialCongestionWindow => {
-                        insert("initial_congestion_window", data.coerce_unsigned().into());
-                    }
-                    ServerMessage2Attributes::QuicTimeout => {
-                        insert("timeout", data.coerce_unsigned().into());
-                    }
-                    // attributes not forming part of the configuration:
-                    ServerMessage2Attributes::WarningMessage
-                    | ServerMessage2Attributes::Invalid => {}
-                }
-            } else {
-                return Err(figment::Error::from(format!(
-                    "Unknown ServerMessage2Attributes tag {}",
-                    attr.tag_raw()
-                )));
-            }
-        }
-
-        let mut profile_map = Map::new();
-        let _ = profile_map.insert(Profile::Global, dict);
-
-        Ok(profile_map)
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////
-// CLOSEDOWN REPORT
-
-#[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
-/// The statistics sent by the server when the job is done
-pub enum ClosedownReport {
-    /// Special value that should never be seen on the wire
-    #[serde(skip_serializing)]
-    Unknown,
-    /// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-    /// On the wire enum discriminant: 1
-    V1(ClosedownReportV1),
-}
-impl ProtocolMessage for ClosedownReport {}
-
-/// Version 1 of the closedown report.
-/// This version was introduced in qcp 0.3 with `VersionCompatibility` level 1.
-#[derive(Serialize, Deserialize, PartialEq, Default, Clone, derive_more::Debug)]
-pub struct ClosedownReportV1 {
-    /// Final congestion window
-    #[debug("{}", cwnd.0)]
-    pub cwnd: Uint,
-    /// Number of packets sent
-    #[debug("{}", sent_packets.0)]
-    pub sent_packets: Uint,
-    /// Number of packets lost
-    #[debug("{}", lost_packets.0)]
-    pub lost_packets: Uint,
-    /// Number of bytes lost
-    #[debug("{}", lost_bytes.0)]
-    pub lost_bytes: Uint,
-    /// Number of congestion events detected
-    #[debug("{}", congestion_events.0)]
-    pub congestion_events: Uint,
-    /// Number of black holes detected
-    #[debug("{}", black_holes.0)]
-    pub black_holes: Uint,
-    /// Number of bytes sent
-    #[debug("{}", sent_bytes.0)]
-    pub sent_bytes: Uint,
-
-    /// Optional extended data
-    ///
-    /// If it is mandatory for the client to action a given attribute, it MUST NOT be sent in this field.
-    /// Instead, use a later version of the `ClosedownReport`.
-    ///
-    /// This field was added in qcp 0.5 with `VersionCompatibility` level 2.
-    /// Prior to Compatibility::Level(2) this was a reserved u8, which was required to be set to 0.
-    /// If length 0, it looks the same on the wire.
-    /// If length >0, earlier versions ignore the attributes.
-    pub extension: Vec<TaggedData<ClosedownReportExtension>>,
-}
-
-impl From<&ConnectionStats> for ClosedownReportV1 {
-    fn from(stats: &ConnectionStats) -> Self {
-        let ps = &stats.path;
-        // look, nobody will overrun u64 micros except on interstellar connections (but if so, they won't be using qcp)
-        let rtt: u64 = ps.rtt.as_micros().try_into().unwrap_or(u64::MAX);
-        let mut extension = vec![];
-        if rtt != 0 {
-            extension.push(ClosedownReportExtension::Rtt.with_unsigned(rtt));
-        }
-        if ps.current_mtu != 0 {
-            extension.push(ClosedownReportExtension::Pmtu.with_unsigned(ps.current_mtu));
-        }
-
-        Self {
-            cwnd: Uint(ps.cwnd),
-            sent_packets: Uint(ps.sent_packets),
-            sent_bytes: Uint(stats.udp_tx.bytes),
-            lost_packets: Uint(ps.lost_packets),
-            lost_bytes: Uint(ps.lost_bytes),
-            congestion_events: Uint(ps.congestion_events),
-            black_holes: Uint(ps.black_holes_detected),
-            extension,
-        }
-    }
-}
-
-/// Extension attributes for the closedown report
-///
-/// This enum was introduced in qcp 0.5 with `VersionCompatibility` level 2.
-#[derive(strum_macros::Display, Clone, Copy, Debug, IntEnum, PartialEq)]
-#[non_exhaustive]
-#[repr(u64)]
-pub enum ClosedownReportExtension {
-    /// Indicates an invalid attribute.
-    Invalid = 0,
-    /// The Path MTU for the connection, as measured by the server, in bytes
-    Pmtu = 1,
-    /// The Round Trip Time for the connection, as measured by the server, in microseconds
-    Rtt = 2,
-}
-impl DataTag for ClosedownReportExtension {}
-
 // //////////////////////////////////////////////////////////////////////////////////////////////
 
 #[cfg(test)]
@@ -1291,39 +331,29 @@ impl DataTag for ClosedownReportExtension {}
 mod test {
     use std::{
         io::Cursor,
-        net::{IpAddr, Ipv4Addr, Ipv6Addr},
+        net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr},
     };
 
-    use assertables::{assert_contains, assert_matches};
-    use pretty_assertions::{assert_eq, assert_str_eq};
-    use quinn::ConnectionStats;
+    use pretty_assertions::assert_eq;
     use serde::{Deserialize, Serialize};
-    use serde_bare::Uint;
 
     use crate::{
-        config::{Configuration, Configuration_Optional, Manager},
         protocol::{
-            DataTag, FindTag, TaggedData,
             common::ProtocolMessage,
-            control::{
-                ClientMessage2Attributes, ClientMessageAttributes, ClientMessageV1,
-                ClientMessageV2, ClosedownReport, Compatibility, CongestionController,
-                ConnectionType, CredentialsType, Direction, OriginalClientMessage,
-                OriginalClientMessageV1, ServerGreeting, ServerMessage2Attributes, ServerMessageV1,
-                ServerMessageV2,
-            },
-            display_vec_td,
+            control::{Compatibility, CongestionController, ConnectionType},
         },
         util::{PortRange as CliPortRange, serialization::SerializeAsString},
     };
 
-    use super::{
-        ClientGreeting, ClientMessage, ClosedownReportV1, PortRange_OnWire, ServerFailure,
-        ServerMessage,
-    };
+    use super::PortRange_OnWire;
+
+    // helper function - creates a bogus certificate
+    pub(crate) fn dummy_cert() -> Vec<u8> {
+        vec![0, 1, 2]
+    }
 
     #[test]
-    fn test_connection_type() {
+    fn convert_connection_type() {
         let ip4 = IpAddr::from(Ipv4Addr::LOCALHOST);
         let ct4 = ConnectionType::from(ip4);
         assert_eq!(ct4, ConnectionType::Ipv4);
@@ -1331,24 +361,14 @@ mod test {
         let ip6 = IpAddr::from(Ipv6Addr::LOCALHOST);
         let ct6 = ConnectionType::from(ip6);
         assert_eq!(ct6, ConnectionType::Ipv6);
-    }
 
-    #[test]
-    fn test_closedown_report() {
-        use serde_bare::Uint;
+        let sa4: SocketAddr = "127.0.0.1:1234".parse().unwrap();
+        let ct4 = ConnectionType::from(sa4);
+        assert_eq!(ct4, ConnectionType::Ipv4);
 
-        let mut stats = ConnectionStats::default();
-        stats.path.cwnd = 42;
-        stats.path.black_holes_detected = 88;
-        stats.udp_tx.bytes = 12345;
-        let report = ClosedownReportV1::from(&stats);
-        let expected = ClosedownReportV1 {
-            cwnd: Uint(42),
-            black_holes: Uint(88),
-            sent_bytes: Uint(12345),
-            ..Default::default()
-        };
-        assert_eq!(report, expected);
+        let sa6: SocketAddr = "[::1]:4321".parse().unwrap();
+        let ct6 = ConnectionType::from(sa6);
+        assert_eq!(ct6, ConnectionType::Ipv6);
     }
 
     /// Time-travelling compatibility: Version 1 of the structure.
@@ -1404,198 +424,6 @@ mod test {
     }
 
     #[test]
-    fn serialize_client_greeting() {
-        let msg = ClientGreeting {
-            compatibility: 1,
-            debug: false,
-            extension: 0,
-        };
-        let wire = msg.to_vec().unwrap();
-        let deser = ClientGreeting::from_slice(&wire).unwrap();
-        assert_eq!(msg, deser);
-    }
-
-    #[test]
-    fn serialize_server_greeting() {
-        let msg = ServerGreeting {
-            compatibility: 1,
-            extension: 0,
-        };
-        let wire = msg.to_vec().unwrap();
-        let deser = ServerGreeting::from_slice(&wire).unwrap();
-        assert_eq!(msg, deser);
-    }
-
-    fn dummy_cert() -> Vec<u8> {
-        vec![0, 1, 2]
-    }
-
-    #[test]
-    fn serialize_client_message() {
-        let config = Configuration_Optional {
-            tx: Some(42u64.into()),
-            rx: Some(89u64.into()),
-            rtt: Some(1234),
-            congestion: Some(CongestionController::Bbr.into()),
-            udp_buffer: Some(456_789u64.into()),
-            initial_congestion_window: Some(12345u64.into()),
-            port: Some(CliPortRange { begin: 17, end: 98 }),
-            remote_port: Some(CliPortRange {
-                begin: 123,
-                end: 456,
-            }),
-            remote_user: None,
-            timeout: Some(432),
-            // other client options are irrelevant to this test but we'll specify them anyway so we can rely on the compiler to catch any missing fields
-            packet_threshold: None,
-            time_threshold: None,
-            initial_mtu: None,
-            min_mtu: None,
-            max_mtu: None,
-
-            address_family: None,
-            ssh: None,
-            ssh_options: None,
-            time_format: None,
-            ssh_config: None,
-            ssh_subsystem: None,
-            color: None,
-            tls_auth_type: None,
-        };
-
-        let cmsg = {
-            let cert = CredentialsType::X509.with_variant(dummy_cert().into());
-            let mut manager = Manager::without_default(None);
-            manager.merge_provider(&config);
-            let cfg = manager.get::<Configuration_Optional>().unwrap();
-            ClientMessage::new(
-                Compatibility::Level(1),
-                cert,
-                ConnectionType::Ipv4,
-                false,
-                &cfg,
-            )
-        };
-        let ser = cmsg.to_vec().unwrap();
-        //println!("{cmsg:#?}");
-        //println!("vec: {ser:?}");
-        let deser = ClientMessage::from_slice(&ser).unwrap();
-        //println!("{deser:#?}");
-
-        let disp = format!("{cmsg}");
-        eprintln!("{disp}");
-        assert!(disp.contains("123-456"));
-
-        let _empty: Vec<TaggedData<ClientMessageAttributes>> = vec![];
-        assert_matches!(
-            deser,
-            ClientMessage::V1(ClientMessageV1 {
-                cert: _,
-                connection_type: ConnectionType::Ipv4,
-                port: Some(PortRange_OnWire {
-                    // crucial check: this is client config.remote_port
-                    begin: 123,
-                    end: 456
-                }),
-                show_config: false,
-                bandwidth_to_server: Some(Uint(42)),
-                bandwidth_to_client: Some(Uint(89)),
-                rtt: Some(1234),
-                congestion: Some(CongestionController::Bbr),
-                initial_congestion_window: Some(Uint(12345)),
-                timeout: Some(432),
-                attributes: _empty,
-            })
-        );
-    }
-
-    #[test]
-    fn construct_client_message() {
-        // additional serialization cases not tested by serialize_and_provide_client_message
-        let cert = CredentialsType::X509.with_variant(dummy_cert().into());
-        let mut manager = Manager::without_default(None);
-        let config = Configuration_Optional::default();
-        manager.merge_provider(&config);
-        let cfg = manager.get::<Configuration_Optional>().unwrap();
-        let cmsg = ClientMessage::new(
-            Compatibility::Level(1),
-            cert,
-            ConnectionType::Ipv4,
-            false,
-            &cfg,
-        );
-        assert_matches!(
-            cmsg,
-            ClientMessage::V1(ClientMessageV1 {
-                bandwidth_to_server: None,
-                ..
-            })
-        );
-    }
-
-    #[test]
-    fn serialize_provide_server_message() {
-        use engineering_repr::EngineeringQuantity as EQ;
-
-        let v1 = ServerMessageV1 {
-            port: 12345,
-            cert: vec![9, 8, 7],
-            name: "hello".to_string(),
-            bandwidth_to_client: Uint(123),
-            bandwidth_to_server: Uint(456),
-            rtt: 789,
-            congestion: CongestionController::Bbr,
-            initial_congestion_window: Uint(4321),
-            timeout: 42,
-            warning: String::from("this is a warning"),
-            extension: 0,
-        };
-        let msg = ServerMessage::V1(v1.clone());
-        let wire = msg.to_vec().unwrap();
-        let deser = ServerMessage::from_slice(&wire).unwrap();
-        assert_eq!(msg, deser);
-
-        let mut manager = Manager::without_files(None); // with system defaults
-        manager.merge_provider(&v1);
-        let cfg = manager.get::<Configuration>().unwrap();
-        println!("{cfg:?}");
-        let expected = Configuration {
-            // Server message is processed by the client, so bandwidth_to_client becomes config.rx
-            rx: EQ::<u64>::from(v1.bandwidth_to_client.0),
-            tx: EQ::<u64>::from(v1.bandwidth_to_server.0),
-            rtt: v1.rtt,
-            congestion: v1.congestion.into(),
-            initial_congestion_window: v1.initial_congestion_window.0.into(),
-            timeout: v1.timeout,
-            ..Configuration::system_default().clone()
-        };
-        assert_eq!(cfg, expected);
-    }
-
-    #[test]
-    fn serialize_closedown_report() {
-        let msg = ClosedownReport::V1(ClosedownReportV1 {
-            cwnd: Uint(42),
-            sent_packets: Uint(123),
-            lost_packets: Uint(234),
-            lost_bytes: Uint(456_798),
-            congestion_events: Uint(44),
-            black_holes: Uint(22),
-            sent_bytes: Uint(987_654),
-            extension: vec![],
-        });
-        let wire = msg.to_vec().unwrap();
-        let deser = ClosedownReport::from_slice(&wire).unwrap();
-        assert_eq!(msg, deser);
-    }
-
-    #[test]
-    fn skip_serializing() {
-        let msg = ServerMessage::ToFollow;
-        let _ = msg.to_vec().expect_err("ToFollow cannot be serialized");
-    }
-
-    #[test]
     fn type_conversions_congestion() {
         let c = CongestionController::Cubic;
         let c2 = SerializeAsString::<CongestionController>::from(c);
@@ -1617,18 +445,6 @@ mod test {
     }
 
     #[test]
-    fn display_server_failure() {
-        let sf = ServerFailure::Malformed;
-        assert_eq!(format!("{sf}"), "Malformed");
-        let sf = ServerFailure::NegotiationFailed("hello".to_string());
-        assert_eq!(format!("{sf}"), "Negotiation Failed: hello");
-        let sf = ServerFailure::EndpointFailed("hello".to_string());
-        assert_eq!(format!("{sf}"), "Endpoint Failed: hello");
-        let sf = ServerFailure::Unknown("hello".to_string());
-        assert_eq!(format!("{sf}"), "Unknown error: hello");
-    }
-
-    #[test]
     fn compat_level_from_wire() {
         let cases = &[
             (0u16, Compatibility::Level(0)),
@@ -1637,304 +453,21 @@ mod test {
             (32768, Compatibility::Newer),
             (65535, Compatibility::Newer),
         ];
-        for (wire, expected) in cases {
+        for (wire, compat) in cases {
             let level: Compatibility = (*wire).into();
             assert_eq!(
-                level, *expected,
-                "wire {wire} should be {expected:?} but got {level}"
+                level, *compat,
+                "wire {wire} should be {compat:?} but got {level}"
             );
+            let wire2 = u16::from(*compat);
+            if *compat == Compatibility::Newer {
+                assert_eq!(wire2, 0, "compat Newer should be wire 0");
+            } else {
+                assert_eq!(
+                    wire2, *wire,
+                    "compat {compat:?} failed to convert back (expected {wire})"
+                );
+            }
         }
-    }
-
-    #[test]
-    fn wire_marshalling_client_greeting() {
-        // This message is critical to the entire protocol. It cannot change without breaking compatibility.
-        let msg = ClientGreeting {
-            compatibility: 1,
-            debug: true,
-            extension: 3,
-        };
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x01\x00\x01\x03".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn wire_marshalling_server_greeting() {
-        // This message is critical to the entire protocol. It cannot change without breaking compatibility.
-        let msg = ServerGreeting {
-            compatibility: 1,
-            extension: 4,
-        };
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x01\x00\x04".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn wire_marshalling_client_message_v1() {
-        let cert = dummy_cert();
-        let msg = ClientMessage::V1(ClientMessageV1::new(
-            &cert,
-            ConnectionType::Ipv4,
-            false,
-            &Configuration_Optional::default(),
-        ));
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x01\x03\x00\x01\x02\x04\x00\x00\x00\x00\x00\x00\x00\x00\x00".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn wire_marshalling_server_message_v1() {
-        let msg = ServerMessage::V1(ServerMessageV1 {
-            port: 12345,
-            cert: vec![9, 8, 7],
-            name: "hello".to_string(),
-            bandwidth_to_client: Uint(123),
-            bandwidth_to_server: Uint(456),
-            rtt: 789,
-            congestion: CongestionController::Bbr,
-            initial_congestion_window: Uint(4321),
-            timeout: 42,
-            warning: String::from("this is a warning"),
-            extension: 0,
-        });
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x0190\x03\x09\x08\x07\x05hello\xc8\x03{\x15\x03\x01\xe1!*\x00\x11this is a warning\x00".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn wire_marshalling_server_message_failure() {
-        let msg = ServerMessage::Failure(ServerFailure::NegotiationFailed("hello".to_string()));
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x02\x01\x05hello".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn wire_marshalling_closedown_report_v1() {
-        let msg = ClosedownReport::V1(ClosedownReportV1 {
-            cwnd: Uint(42),
-            sent_packets: Uint(65),
-            lost_packets: Uint(66),
-            lost_bytes: Uint(456_798),
-            congestion_events: Uint(44),
-            black_holes: Uint(49),
-            sent_bytes: Uint(987_654),
-            extension: vec![],
-        });
-        let wire = msg.to_vec().unwrap();
-        let expected = b"\x01*AB\xde\xf0\x1b,1\x86\xa4<\x00".to_vec();
-        assert_eq!(wire, expected);
-    }
-
-    #[test]
-    fn display_clientmessage_attrs() {
-        let d = ClientMessageAttributes::DirectionOfTravel
-            .with_variant(Direction::ClientToServer.into());
-        let cm: ClientMessage = ClientMessage::V1(ClientMessageV1 {
-            attributes: vec![d.clone()],
-            ..Default::default()
-        });
-        // Debug
-        let s = format!("{d:?}");
-        eprintln!("{s}");
-        assert_str_eq!(
-            s,
-            "TaggedData { tag: ClientMessageAttributes::DirectionOfTravel, data: ClientToServer, .. }"
-        );
-        let s = format!("{cm:?}");
-        eprintln!("{s}");
-        assert!(s.contains("ClientMessageAttributes::DirectionOfTravel, data: ClientToServer"));
-
-        // Display
-        let s = display_vec_td(&vec![d.clone()]);
-        eprintln!("{s}");
-        assert_str_eq!(s, "[DirectionOfTravel:ClientToServer]");
-        let s = format!("{d}");
-        eprintln!("{s}");
-        assert_str_eq!(s, "(DirectionOfTravel, ClientToServer)");
-        let s = format!("{cm}");
-        eprintln!("{s}");
-        assert!(s.contains("[DirectionOfTravel:ClientToServer]"));
-    }
-
-    #[test]
-    fn clientmessagev1_attrs_backwards_compat() {
-        let d = ClientMessageAttributes::DirectionOfTravel
-            .with_variant(Direction::ClientToServer.into());
-        let cm = ClientMessage::V1(ClientMessageV1 {
-            attributes: vec![d.clone()],
-            ..Default::default()
-        });
-        let wire = cm.to_vec().unwrap();
-        let decode = OriginalClientMessage::from_slice(&wire).unwrap();
-        // This is really a no-crash test.
-        assert_eq!(
-            decode,
-            OriginalClientMessage::V1(OriginalClientMessageV1 {
-                cert: vec![],
-                connection_type: ConnectionType::Ipv4,
-                port: None,
-                show_config: false,
-                bandwidth_to_server: None,
-                bandwidth_to_client: None,
-                rtt: None,
-                congestion: None,
-                initial_congestion_window: None,
-                timeout: None,
-                extension: 1, // Earlier versions ignore this field, so if the assert passes we're good.
-            })
-        );
-    }
-
-    #[test]
-    fn client_message_2_debug_attrs() {
-        let msg = ClientMessageV2 {
-            attributes: vec![
-                ClientMessage2Attributes::Invalid.into(),
-                ClientMessage2Attributes::DirectionOfTravel
-                    .with_unsigned(Direction::ClientToServer as u64),
-                ClientMessage2Attributes::CongestionControllerType
-                    .with_unsigned(CongestionController::NewReno as u64),
-                ClientMessage2Attributes::OutputConfig.into(),
-            ],
-            ..Default::default()
-        };
-        let s = format!("{msg:?}");
-        //eprintln!("{s}");
-        assert!(s.contains("Invalid, data: Empty"));
-        assert!(s.contains("DirectionOfTravel, data: ClientToServer"));
-        assert!(s.contains("CongestionControllerType, data: newreno"));
-        assert!(s.contains("OutputConfig, data: Empty"));
-    }
-
-    #[test]
-    fn client_message_2_display() {
-        let msg = ClientMessageV2 {
-            attributes: vec![
-                ClientMessage2Attributes::BandwidthToClient.with_unsigned(123_456_789u32),
-                ClientMessage2Attributes::BandwidthToServer.with_unsigned(32_768u32),
-                ClientMessage2Attributes::Invalid.into(),
-                ClientMessage2Attributes::DirectionOfTravel
-                    .with_unsigned(Direction::ClientToServer as u64),
-                ClientMessage2Attributes::CongestionControllerType
-                    .with_unsigned(CongestionController::NewReno as u64),
-                ClientMessage2Attributes::OutputConfig.into(),
-            ],
-            ..Default::default()
-        };
-        let s = format!("{msg}");
-        eprintln!("{s}");
-        assert_contains!(s, "Invalid:Empty");
-        assert_contains!(s, "BandwidthToClient:123.4M");
-        assert_contains!(s, "BandwidthToServer:32.76k");
-        assert_contains!(s, "DirectionOfTravel:ClientToServer");
-        assert_contains!(s, "CongestionControllerType:newreno");
-        assert_contains!(s, "OutputConfig:Empty");
-    }
-
-    #[test]
-    fn server_message_1_2_conversion() {
-        use FindTag as _;
-
-        let mut msg1 = ServerMessageV1::default();
-        msg1.warning.push('a');
-        msg1.initial_congestion_window.0 = 42;
-        msg1.timeout = 7;
-
-        let msg2 = ServerMessageV2::from(msg1);
-        let attrs = &msg2.attributes;
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::WarningMessage)
-            .unwrap();
-        assert_eq!(tag.as_str(), Some("a"));
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::InitialCongestionWindow)
-            .unwrap();
-        assert_eq!(tag.coerce_unsigned(), 42);
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::QuicTimeout)
-            .unwrap();
-        assert_eq!(tag.coerce_unsigned(), 7);
-    }
-
-    #[test]
-    fn server_message_2_config_attrs() {
-        let mut mgr = Manager::without_files(None);
-        let cfg = Configuration_Optional {
-            congestion: Some(CongestionController::Bbr.into()),
-            initial_congestion_window: Some(42u64.into()),
-            timeout: Some(88),
-            ..Default::default()
-        };
-        mgr.merge_provider(&cfg);
-        let final_cfg = mgr.get::<Configuration>().unwrap();
-        let mut msg = ServerMessageV2::default();
-        msg.apply_config_attributes(&final_cfg);
-
-        let attrs = &msg.attributes;
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::CongestionController)
-            .unwrap();
-        assert_eq!(tag.coerce_unsigned(), CongestionController::Bbr as u64);
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::InitialCongestionWindow)
-            .unwrap();
-        assert_eq!(tag.coerce_unsigned(), 42);
-
-        let tag = attrs
-            .find_tag(ServerMessage2Attributes::QuicTimeout)
-            .unwrap();
-        assert_eq!(tag.coerce_unsigned(), 88);
-    }
-
-    #[test]
-    fn server_message_2_to_provider() {
-        let msg = ServerMessageV2 {
-            bandwidth_to_server: Uint(12345),
-            bandwidth_to_client: Uint(54321),
-            rtt: 42,
-            attributes: vec![
-                ServerMessage2Attributes::CongestionController
-                    .with_unsigned(CongestionController::Bbr as u64),
-                ServerMessage2Attributes::InitialCongestionWindow.with_unsigned(5544u32),
-                ServerMessage2Attributes::QuicTimeout.with_unsigned(55u32),
-                // these two are not part of the config:
-                ServerMessage2Attributes::WarningMessage.with_variant("hi".into()),
-                ServerMessage2Attributes::Invalid.into(),
-            ],
-            ..Default::default()
-        };
-        let mut mgr = Manager::without_files(None);
-        mgr.apply_system_default();
-        mgr.merge_provider(msg);
-        let cfg = mgr.get::<Configuration>().unwrap();
-        assert_eq!(cfg.rx, 54321u64.into());
-        assert_eq!(cfg.tx, 12345u64.into());
-        assert_eq!(cfg.rtt, 42);
-        assert_eq!(cfg.congestion, CongestionController::Bbr.into());
-        assert_eq!(cfg.initial_congestion_window, 5544u32.into());
-        assert_eq!(cfg.timeout, 55);
-    }
-
-    #[test]
-    fn server_message_2_display() {
-        let msg = ServerMessageV2 {
-            bandwidth_to_client: Uint(12_000),
-            bandwidth_to_server: Uint(1_987_654_321),
-            ..Default::default()
-        };
-        let s = format!("{msg}");
-        eprintln!("{s}");
-        assert_contains!(s, "in 1.987G");
-        assert_contains!(s, "out 12k");
     }
 }
